requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aio-pika==9.5.5
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiormq==6.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.3.0
bert-score==0.3.13
bitsandbytes==0.42.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
codecarbon==2.8.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dm-tree==0.1.9
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.6
fief-client==0.20.0
filelock==3.16.1
Flask==3.1.0
Flask-API==3.1
flatbuffers==25.2.10
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gritql==0.1.5
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jwcrypto==1.5.6
kagglehub==0.3.11
keras==3.9.0
kiwisolver==1.4.8
langchain==0.3.14
langchain-cli==0.0.35
langchain-community==0.3.14
langchain-core==0.3.29
langchain-ollama==0.2.2
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langserve==0.3.1
langsmith==0.2.10
libclang==18.1.1
logger==1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
mdurl==0.1.2
metrics==0.0.2
mistralai==1.2.6
ml_dtypes==0.5.1
mlcommons_loadgen==5.0.15
mpmath==1.3.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-ml-py==12.570.86
ollama==0.4.5
onnxruntime==1.21.0
openai==1.59.6
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.14
packaging==24.2
pamqp==3.3.0
pandas==2.2.3
pillow==11.1.0
priority==2.0.0
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.21.12
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
Quart==0.20.0
questionary==2.1.0
RapidFuzz==3.12.2
redis==5.2.0
fakeredis==2.26.1
redis-sentinel-connector==0.1.8
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rouge-score==0.1.2
safetensors==0.5.2
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
sse-starlette==1.8.2
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
taskgroup==0.2.2
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-datasets==4.9.8
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.0
termcolor==2.3.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.0
typer==0.9.4
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
yaspin==3.1.0
zipp==3.21.0
